dev: bugs fixed
