+++ css/style.css
@@ -1,0 +1,44 @@
+body {
+    background: linear-gradient(135deg, #f0f0f0, #d9d9d9);
+    color: #333;
+    font-family: 'Helvetica Neue', sans-serif;
+}
+.main {
+    padding: 2rem;
+    border-radius: 15px;
+    box-shadow: 0 4px 30px rgba(0, 0, 0, 0.2);
+    background-color: black;
+    transition: all 0.5s ease;
+}
+h1 {
+    text-align: center;
+    font-size: 3rem;
+    color: #007aff;
+    text-shadow: 1px 1px #e3e3e3;
+}
+h2 {
+    text-align: center;
+    font-size: 2rem;
+    color: #333;
+}
+.tagline {
+    text-align: center;
+    font-size: 1.5rem;
+    margin-bottom: 20px;
+}
+.stButton button {
+    background-color: #007aff;
+    color: white;
+    border: none;
+    padding: 12px 30px;
+    border-radius: 5px;
+    transition: background-color 0.3s ease-in-out, transform 0.2s ease-in-out;
+}
+.stButton button:hover {
+    background-color: #005bb5;
+    transform: scale(1.05);
+}
+.footer {
+    text-align: center;
+    margin-top: 20px;
+}
